wrap the rule into rulepushdownfilterjoin
