+++ src/query/service/src/sql/planner/optimizer/heuristic/heuristic.rs
@@ -47,7 +47,6 @@ pub static DEFAULT_REWRITE_RULES: Lazy<Vec<RuleID>> = Lazy::new(|| {
         RuleID::FoldCountAggregate,
         RuleID::SplitAggregate,
         RuleID::PushDownFilterScan,
-        RuleID::EliminateOuterJoin,
     ]
 });
