+++ src/query/service/src/sql/planner/optimizer/rule/rewrite/mod.rs
@@ -14,7 +14,6 @@
 
 mod rule_eliminate_eval_scalar;
 mod rule_eliminate_filter;
-mod rule_eliminate_outer_join;
 mod rule_fold_count_aggregate;
 mod rule_merge_eval_scalar;
 mod rule_merge_filter;
@@ -31,7 +30,6 @@ mod rule_split_aggregate;
 
 pub use rule_eliminate_eval_scalar::RuleEliminateEvalScalar;
 pub use rule_eliminate_filter::RuleEliminateFilter;
-pub use rule_eliminate_outer_join::RuleEliminateOuterJoin;
 pub use rule_fold_count_aggregate::RuleFoldCountAggregate;
 pub use rule_merge_eval_scalar::RuleMergeEvalScalar;
 pub use rule_merge_filter::RuleMergeFilter;
