+++ src/query/service/src/sql/planner/optimizer/rule/rewrite/rule_push_down_filter_join.rs
@@ -14,6 +14,8 @@
 
 use common_datavalues::type_coercion::compare_coercion;
 use common_exception::Result;
+use common_planner::ColumnSet;
+use common_planner::IndexType;
 
 use crate::sql::binder::wrap_cast;
 use crate::sql::binder::JoinPredicate;
@@ -27,6 +29,7 @@ use crate::sql::plans::JoinType;
 use crate::sql::plans::LogicalInnerJoin;
 use crate::sql::plans::PatternPlan;
 use crate::sql::plans::RelOp;
+use crate::sql::plans::Scalar;
 use crate::sql::plans::ScalarExpr;
 
 pub struct RulePushDownFilterJoin {
@@ -70,6 +73,150 @@ impl RulePushDownFilterJoin {
             ),
         }
     }
+
+    #[allow(clippy::only_used_in_recursion)]
+    fn find_nullable_columns(
+        &self,
+        predicate: &Scalar,
+        left_output_columns: &ColumnSet,
+        right_output_columns: &ColumnSet,
+        nullable_columns: &mut Vec<IndexType>,
+    ) -> Result<()> {
+        match predicate {
+            Scalar::BoundColumnRef(column_binding) => {
+                nullable_columns.push(column_binding.column.index);
+            }
+            Scalar::OrExpr(expr) => {
+                let mut left_cols = vec![];
+                let mut right_cols = vec![];
+                self.find_nullable_columns(
+                    &expr.left,
+                    left_output_columns,
+                    right_output_columns,
+                    &mut left_cols,
+                )?;
+                self.find_nullable_columns(
+                    &expr.right,
+                    left_output_columns,
+                    right_output_columns,
+                    &mut right_cols,
+                )?;
+                if !left_cols.is_empty() && !right_cols.is_empty() {
+                    for left_col in left_cols.iter() {
+                        for right_col in right_cols.iter() {
+                            if (left_output_columns.contains(left_col)
+                                && left_output_columns.contains(right_col))
+                                || (right_output_columns.contains(left_col)
+                                    && right_output_columns.contains(right_col))
+                            {
+                                nullable_columns.push(*left_col);
+                                break;
+                            }
+                        }
+                    }
+                }
+            }
+            Scalar::ComparisonExpr(expr) => {
+                // For any comparison expr, if input is null, the compare result is false
+                self.find_nullable_columns(
+                    &expr.left,
+                    left_output_columns,
+                    right_output_columns,
+                    nullable_columns,
+                )?;
+                self.find_nullable_columns(
+                    &expr.right,
+                    left_output_columns,
+                    right_output_columns,
+                    nullable_columns,
+                )?;
+            }
+            Scalar::CastExpr(expr) => {
+                self.find_nullable_columns(
+                    &expr.argument,
+                    left_output_columns,
+                    right_output_columns,
+                    nullable_columns,
+                )?;
+            }
+            // `predicate` can't be `Scalar::AndExpr`
+            // because `Scalar::AndExpr` had been split in binder
+            _ => {}
+        }
+        Ok(())
+    }
+
+    fn convert_outer_to_inner_join(&self, s_expr: &SExpr) -> Result<SExpr> {
+        let filter: Filter = s_expr.plan().clone().try_into()?;
+        let mut join: LogicalInnerJoin = s_expr.child(0)?.plan().clone().try_into()?;
+        let origin_join_type = join.join_type.clone();
+        let s_join_expr = s_expr.child(0)?;
+        let join_expr = RelExpr::with_s_expr(s_join_expr);
+        let left_child_output_column = join_expr.derive_relational_prop_child(0)?.output_columns;
+        let right_child_output_column = join_expr.derive_relational_prop_child(1)?.output_columns;
+        let predicates = &filter.predicates;
+        let mut nullable_columns: Vec<IndexType> = vec![];
+        for predicate in predicates {
+            self.find_nullable_columns(
+                predicate,
+                &left_child_output_column,
+                &right_child_output_column,
+                &mut nullable_columns,
+            )?;
+        }
+
+        if join.join_type == JoinType::Left
+            || join.join_type == JoinType::Right
+            || join.join_type == JoinType::Full
+        {
+            let mut left_join = false;
+            let mut right_join = false;
+            for col in nullable_columns.iter() {
+                if left_child_output_column.contains(col) {
+                    right_join = true;
+                }
+                if right_child_output_column.contains(col) {
+                    left_join = true;
+                }
+            }
+
+            match join.join_type {
+                JoinType::Left => {
+                    if left_join {
+                        join.join_type = JoinType::Inner
+                    }
+                }
+                JoinType::Right => {
+                    if right_join {
+                        join.join_type = JoinType::Inner
+                    }
+                }
+                JoinType::Full => {
+                    if left_join && right_join {
+                        join.join_type = JoinType::Inner
+                    } else if left_join {
+                        join.join_type = JoinType::Right
+                    } else if right_join {
+                        join.join_type = JoinType::Left
+                    }
+                }
+                _ => unreachable!(),
+            }
+        }
+
+        let changed_join_type = join.join_type.clone();
+        if origin_join_type == changed_join_type {
+            return Ok(s_expr.clone());
+        }
+        let mut result = SExpr::create_binary(
+            join.into(),
+            s_join_expr.child(0)?.clone(),
+            s_join_expr.child(1)?.clone(),
+        );
+        // wrap filter s_expr
+        result = SExpr::create_unary(filter.into(), result);
+        Ok(result)
+    }
 }
 
 impl Rule for RulePushDownFilterJoin {
@@ -78,6 +225,8 @@ impl Rule for RulePushDownFilterJoin {
     }
 
     fn apply(&self, s_expr: &SExpr, state: &mut TransformState) -> Result<()> {
+        // First, try to convert outer join to inner join
+        let s_expr = self.convert_outer_to_inner_join(s_expr)?;
         let filter: Filter = s_expr.plan().clone().try_into()?;
         let join_expr = s_expr.child(0)?;
         let mut join: LogicalInnerJoin = join_expr.plan().clone().try_into()?;
